requirements.txt
torch
torchvision
scikit-learn
pillow
numpy
pandas
streamlit
wikipedia
gTTS
joblib
kaggle
